Extract v2 eval condition helpers
